Better formatting for the README.md

+++ part4/blog/controllers/blogs.js
@@ -51,8 +51,18 @@ blogRouter.post('/', async (request, response) => {
 
 blogRouter.delete('/:id', async(request,response,next) => {
   try{
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token ,process.env.SECRET)
+    if(!decodedToken.id) return response.status(401).json({ error: 'token invalid' })
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if(!blog) return response.status(404).json({ error: 'blog not found' })
+
+    if (user.id.toString() === blog.users.toString()){
+      await Blog.findByIdAndDelete(request.params.id)
+
+      response.status(204).end()
+    }
   }catch (error){
     next(error)
   }
